backend/pyproject.toml
[project]
name = "atlas-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
    "pymongo>=4.6.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "yfinance>=0.2.0",
    "pandas>=2.0.0",
    "alpaca-py>=0.28.0",
    "langgraph>=0.2.0",
    "atlas-agents",
]

[tool.uv.sources]
atlas-agents = { path = "../agents", editable = true }

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
